requirements.txt
# BuildHelper python dependency list
gitpython
PyYAML

